[object] Refactor ObjectRuntime split runtime.rs to mods
